+++ uCNC/src/hal/mcus/samd21/mcu_samd21.c
@@ -710,9 +710,6 @@ bool mcu_rx_ready(void)
 #ifndef mcu_putc
 void mcu_putc(char c)
 {
-#if !(ACTIVITY_LED < 0)
-	mcu_toggle_output(ACTIVITY_LED);
-#endif
 #if (INTERFACE == INTERFACE_USB)
 	if (c != 0)
 	{
@@ -744,9 +741,6 @@ void mcu_putc(char c)
 #ifndef mcu_getc
 char mcu_getc(void)
 {
-#if !(ACTIVITY_LED < 0)
-	mcu_toggle_output(ACTIVITY_LED);
-#endif
 #if (INTERFACE == INTERFACE_USB)
 	while (!tud_cdc_available())
 	{